img resize, compress, convert util func is done
